Read all directory entries before we start invoking the callbacks
Most nfs servers return READDIRPLUS entry cookies as the offset into the directory where the dirent starts.
This means that reading the directory is not impacted by changes to the directory and that BAD_COOKIE errors will never arise.
Think loops like
 opendir()
 while(readdir()){
     unlink()
 }
 closedir()

Some nfs servers can not easily produce semantics similar to the above in which case the cleanup funciton that does a recursive delete using a loop semantically similar to the above.
Change how we traverse directories to instead work something like
 opendir()
 while(readdir()){
    store dirent in list
 }
 closedir()
 walk whole list {
   unlink()
 }