loadfile: add a check for the loadfile option
The loadfile option is currently used without any check
causing segmentation faults. In the past, there was an
initial default value that solved this problem.
The same was true for backend:

struct options options = {
        .backend             = "fileio",
        .loadfile            = DATADIR "/client.txt",
		....

Although I liked the initial default values, it seems that
backend is now using a check instead. I am not sure what
the reasons are, but to follow the same approach here I
I did the same for loadfile.

Signed-off-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>