dbench: assume the libraries are not ready for cross
When using a cross-compiler the current configure script
was assuming that the cross-compiler had all of the optional
libraries without properly checking:
	ac_cv_have_libsmbclient=yes
	ac_cv_have_libiscsi=yes
	ac_cv_linux_scsi_sg=yes
	ac_cv_have_libnfs=yes

Set them to "no" by default. This way, if the cross-compiler
does not have them they will be configured out and we will
be able to build dbench. On the other hand, if the cross-compiler
does have them, from my tests, configure is actually smart enough
to actually enable the cache variables (ac_cv_xx).

Signed-off-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>